chore(auth-data): remove remote module
