+++ feature/auth/data/remote-firebase/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/remote/firebase/service/register/RegisterService.kt
@@ -1,8 +1,8 @@
 package ch.app.hk.bank.locator.feature.auth.data.remote.firebase.service.register
 
 import ch.app.hk.bank.locator.core.threading.DispatcherIo
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.datasource.RegisterRemoteDataSource
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.response.RegisterResponse
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.datasource.RegisterRemoteDataSource
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.response.RegisterResponse
 import ch.app.library.hiltwrap.annotation.HiltWrapBindModule
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseAuthException
