+++ feature/auth/data/remote-firebase/src/test/kotlin/ch/app/hk/bank/locator/feature/auth/data/remote/firebase/service/login/LoginServiceTest.kt
@@ -1,6 +1,6 @@
 package ch.app.hk.bank.locator.feature.auth.data.remote.firebase.service.login
 
-import ch.app.hk.bank.locator.feature.auth.data.remote.login.response.LoginResponse
+import ch.app.hk.bank.locator.feature.auth.data.repo.login.remote.response.LoginResponse
 import ch.app.hk.bank.locator.testing.google.play.services.task.mockTaskError
 import ch.app.hk.bank.locator.testing.google.play.services.task.mockTaskResult
 import com.google.firebase.FirebaseTooManyRequestsException
