--- feature/auth/data/remote/build.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-    alias(libs.plugins.app.jvm)
-}
