--- feature/auth/data/remote/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/remote/login/datasource/LoginRemoteDataSource.kt
@@ -1,10 +1,0 @@
-package ch.app.hk.bank.locator.feature.auth.data.remote.login.datasource
-
-import ch.app.hk.bank.locator.feature.auth.data.remote.login.response.LoginResponse
-
-interface LoginRemoteDataSource {
-    suspend fun emailPasswordLogin(
-        email: String,
-        password: String,
-    ): LoginResponse
-}
