--- feature/auth/data/remote/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/remote/register/datasource/RegisterRemoteDataSource.kt
@@ -1,10 +1,0 @@
-package ch.app.hk.bank.locator.feature.auth.data.remote.register.datasource
-
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.response.RegisterResponse
-
-interface RegisterRemoteDataSource {
-    suspend fun emailPasswordRegister(
-        email: String,
-        password: String,
-    ): RegisterResponse
-}
