--- feature/auth/data/remote/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/remote/user/datasource/UserRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package ch.app.hk.bank.locator.feature.auth.data.remote.user.datasource
-
-import ch.app.hk.bank.locator.feature.auth.data.remote.user.response.UserResponse
-
-interface UserRemoteDataSource {
-    suspend fun getCurrentUser(): UserResponse?
-}
