+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/login/remote/datasource/LoginRemoteDataSource.kt
@@ -1,0 +1,10 @@
+package ch.app.hk.bank.locator.feature.auth.data.repo.login.remote.datasource
+
+import ch.app.hk.bank.locator.feature.auth.data.repo.login.remote.response.LoginResponse
+
+interface LoginRemoteDataSource {
+    suspend fun emailPasswordLogin(
+        email: String,
+        password: String,
+    ): LoginResponse
+}
