+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/register/remote/datasource/RegisterRemoteDataSource.kt
@@ -1,0 +1,10 @@
+package ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.datasource
+
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.response.RegisterResponse
+
+interface RegisterRemoteDataSource {
+    suspend fun emailPasswordRegister(
+        email: String,
+        password: String,
+    ): RegisterResponse
+}
