+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/register/remote/response/RegisterResponse.kt
@@ -1,4 +1,4 @@
-package ch.app.hk.bank.locator.feature.auth.data.remote.register.response
+package ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.response
 
 sealed interface RegisterResponse {
     data class Success(
