+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/register/repository/RegisterRepositoryImpl.kt
@@ -1,14 +1,14 @@
 package ch.app.hk.bank.locator.feature.auth.data.repo.register.repository
 
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.datasource.RegisterRemoteDataSource
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.response.RegisterResponse
 import ch.app.hk.bank.locator.feature.auth.data.repo.register.model.RegisterErrorCode
 import ch.app.hk.bank.locator.feature.auth.data.repo.register.model.RegisterResult
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.datasource.RegisterRemoteDataSource
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.response.RegisterResponse
 import ch.app.library.hiltwrap.annotation.HiltWrapBindModule
 import javax.inject.Inject
 
 @HiltWrapBindModule
-class RegisterRepositoryImpl @Inject constructor(
+internal class RegisterRepositoryImpl @Inject constructor(
     private val registerRemoteDataSource: RegisterRemoteDataSource,
 ) : RegisterRepository {
     override suspend fun emailPasswordRegister(
