+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/user/mapper/UserModelMapper.kt
@@ -1,7 +1,7 @@
 package ch.app.hk.bank.locator.feature.auth.data.repo.user.mapper
 
-import ch.app.hk.bank.locator.feature.auth.data.remote.user.response.UserResponse
 import ch.app.hk.bank.locator.feature.auth.data.repo.user.model.UserModel
+import ch.app.hk.bank.locator.feature.auth.data.repo.user.remote.response.UserResponse
 import org.mapstruct.Mapper
 import org.mapstruct.Mapping
