+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/user/remote/datasource/UserRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package ch.app.hk.bank.locator.feature.auth.data.repo.user.remote.datasource
+
+import ch.app.hk.bank.locator.feature.auth.data.repo.user.remote.response.UserResponse
+
+interface UserRemoteDataSource {
+    suspend fun getCurrentUser(): UserResponse?
+}
