+++ feature/auth/data/repo/src/main/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/user/remote/response/UserResponse.kt
@@ -1,4 +1,4 @@
-package ch.app.hk.bank.locator.feature.auth.data.remote.user.response
+package ch.app.hk.bank.locator.feature.auth.data.repo.user.remote.response
 
 data class UserResponse(
     val displayName: String,
