+++ feature/auth/data/repo/src/test/kotlin/ch/app/hk/bank/locator/feature/auth/data/repo/register/repository/RegisterRepositoryImplTest.kt
@@ -1,9 +1,9 @@
 package ch.app.hk.bank.locator.feature.auth.data.repo.register.repository
 
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.datasource.RegisterRemoteDataSource
-import ch.app.hk.bank.locator.feature.auth.data.remote.register.response.RegisterResponse
 import ch.app.hk.bank.locator.feature.auth.data.repo.register.model.RegisterErrorCode
 import ch.app.hk.bank.locator.feature.auth.data.repo.register.model.RegisterResult
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.datasource.RegisterRemoteDataSource
+import ch.app.hk.bank.locator.feature.auth.data.repo.register.remote.response.RegisterResponse
 import io.kotest.matchers.shouldBe
 import io.mockk.coEvery
 import io.mockk.mockk
